+++ apps/site/types/author.ts
@@ -1,0 +1,12 @@
+export interface AuthorProps {
+  names?: Array<string>;
+  usernames?: Array<string>;
+  clickable?: boolean;
+  container?: HTMLElement | null;
+}
+
+export interface Author {
+  id: string;
+  name: string;
+  website?: string;
+}
