+++ apps/site/types/config.ts
@@ -29,9 +29,3 @@ export interface SiteConfig {
   websiteBanners: Record<string, WebsiteBanner>;
   websiteBadges: Record<string, WebsiteBadge>;
 }
-
-export interface Author {
-  id: string;
-  name: string;
-  website?: string;
-}
