+++ apps/site/types/index.ts
@@ -11,3 +11,4 @@ export * from './server';
 export * from './github';
 export * from './calendar';
 export * from './search';
+export * from './author';
